requirements.txt
flask==2.3.3
python-dotenv==1.0.0
psycopg2-binary==2.9.7
langchain==0.0.267
langchain-experimental==0.0.10
langchain-community==0.0.6
langchain-google-genai==0.0.3
google-generativeai==0.3.1
SQLAlchemy==2.0.20